Get rid of SizeTree subclasses
